python-dspy-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
dspy-ai==2.4.0
openai==1.3.0
requests==2.31.0
